refactor: move tree related function to tree module, add input path normalization
