+++ src/commands/zip.ts
@@ -1,4 +1,5 @@
 import { is_windows, preset_compression_level } from '@/core/constants';
+import { printfile_list_as_file_tree } from '@/core/tree';
 import { list_entries } from '@/core/walk';
 import { create_zip } from '@/core/zip';
 import { logger } from '@/logger';
@@ -7,7 +8,7 @@ import { confirm_overwrite_prompt } from '@/prompts/confirm-overwrite';
 import { createCommand } from '@/utils/command';
 import { log_conflicts } from '@/utils/conflicts';
 import { exists, unique_entries } from '@/utils/fs';
-import { printfile_list_as_file_tree } from '@/utils/path';
+import { normalize_entries } from '@/utils/path';
 import {
   exit_fail_on_error,
   exit_on_finish,
@@ -75,7 +76,7 @@ const zipCommand = createCommand(name, description)
 
 zipCommand.action(async (options) => {
   await exit_fail_on_error(async () => {
-    const unique_inputs = unique_entries(options.input);
+    const unique_inputs = unique_entries(normalize_entries(options.input));
 
     await validation_spinner({
       name: 'output path',
